+++ src/main/kotlin/io/github/guttenbase/statements/AbstractInsertStatementCreator.kt
@@ -1,10 +1,14 @@
 package io.github.guttenbase.statements
 
+import io.github.guttenbase.mapping.PreparedStatementPlaceholderFactory
+import io.github.guttenbase.meta.ColumnMetaData
 import io.github.guttenbase.meta.TableMetaData
 import io.github.guttenbase.repository.ConnectorRepository
+import io.github.guttenbase.repository.hint
 import java.sql.Connection
 import java.sql.PreparedStatement
 import java.sql.SQLException
+import kotlin.Throws
 
 /**
  * Create INSERT statement with multiple VALUES tuples.
@@ -15,6 +19,8 @@ import java.sql.SQLException
  */
 abstract class AbstractInsertStatementCreator(connectorRepository: ConnectorRepository, targetConnectorId: String) :
   AbstractStatementCreator(connectorRepository, targetConnectorId) {
+  protected val placeholderFactory = connectorRepository.hint<PreparedStatementPlaceholderFactory>(targetConnectorId)
+
   /**
    * Create INSERT statement for the mapped target columns.
    */
@@ -29,14 +35,14 @@ abstract class AbstractInsertStatementCreator(connectorRepository: ConnectorRepo
     val numberOfValuesClauses = if (useMultipleValuesClauses) numberOfRowsPerBatch else 1
     val columns = getMappedTargetColumns(sourceTableMetaData, targetTableMetaData)
     val insert = "INSERT INTO " + targetTableName + " (" + createColumnClause(columns) + ")"
-    val sql = insert + " VALUES\n" + createValueTuples(numberOfValuesClauses, columns.size)
+    val sql = insert + " VALUES\n" + createValueTuples(numberOfValuesClauses, columns)
 
     indicator.debug("Create INSERT statement: $insert")
     return destConnection.prepareStatement(sql)
   }
 
-  protected fun createValueTuples(numberOfValuesClauses: Int, columnCount: Int): String {
-    val tuple = (1..columnCount).joinToString(prefix = "\t(", postfix = ")", transform = { "?" })
+  protected fun createValueTuples(numberOfValuesClauses: Int, columns: List<ColumnMetaData>): String {
+    val tuple = columns.joinToString(prefix = "\t(", postfix = ")", transform = { placeholderFactory.map(it) })
 
     return (1..numberOfValuesClauses).joinToString(separator = ",\n", transform = { tuple })
   }